backend/requirements.txt
# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP 및 파일 처리
requests==2.31.0
python-multipart==0.0.6

# 환경 변수 관리
python-dotenv==1.0.0

# 데이터베이스
psycopg2-binary==2.9.7

# 데이터 검증
pydantic==2.5.0

# 이미지 처리 및 ROI
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0